linux-monitor-backend/pyproject.toml
[project]
name = "linux-monitor-backend"
version = "0.1.0"
description = "Linux系统监控后端服务"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "psutil>=5.9.6",
    "redis>=5.0.1",
    "aioredis>=2.0.1",
    "python-multipart>=0.0.6",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-json-logger>=2.0.7",
    "asyncio-mqtt>=0.16.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
linux-monitor = "app.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.api", "app.core", "app.models", "app.services"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
]
